MS-29 몬스터 에셋 수정

+++ Client/Source/MapleStory/Private/Managers/MsCheatManager.cpp
@@ -38,7 +38,7 @@ void UMsCheatManager::LogMobList() const {
 	
 		const FVector NewLocation = {Location.X - SpawnPointLocation.X, Location.Y, Location.Z - SpawnPointLocation.Z};
 
-		UE_LOG(LogTemp, Warning, TEXT("MobId: %d, X: %f, Y: %f"), Mob->MobId, NewLocation.X, NewLocation.Z);
+		UE_LOG(LogTemp, Warning, TEXT("MobId: %d, X: %d, Y: %d"), Mob->MobId, (int32)NewLocation.X, (int32)NewLocation.Z);
 	}
 }
